# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FERNS — Federated Ecological Resource Network System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /miflora/species:
    get:
      operationId: getMifloraSpecies
      tags: [miflora]
      summary: Look up a species in Michigan Flora
      description: >
        Looks up a vascular plant species in the Michigan Flora REST API.
        Returns the complete passthrough response from Michigan Flora: all search
        records from flora_search_sp, spec_text (taxonomic details and description),
        synonyms, and plant images. Two to four API calls are required per lookup.
        Results are cached permanently (no TTL) — Michigan Flora data does not change;
        use ?refresh=true to force a re-fetch from upstream.
        The pimage_info object in the response is addtively enriched by FERNS with two
        computed fields: image_url and thumbnail_url (constructed from the plant_id and
        image_id using the Michigan Flora static asset URL formula). All other source
        fields are returned unchanged.
        The st field in source records uses the literal string 'NULL' (not JSON null)
        for unknown or absent status — this is a quirk of the Michigan Flora API.
        The c field is always a string; '*' indicates an adventive (non-native) species.
        Scientific names for adventive species are returned ALL-CAPS by the source API.
      parameters:
        - name: name
          in: query
          required: true
          description: Scientific name to look up (e.g. Asclepias tuberosa)
          schema:
            type: string
            minLength: 1
        - name: refresh
          in: query
          required: false
          description: If true, bypasses cache and fetches fresh from Michigan Flora API
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Species lookup complete (found or not found — check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MifloraSpeciesResponse"
        "400":
          description: Missing or empty name parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Michigan Flora API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /miflora/counties:
    get:
      operationId: getMifloraCounties
      tags: [miflora]
      summary: Get county-level occurrence records for a Michigan Flora species by name
      description: >
        Returns county-level occurrence records for all 83 Michigan counties for
        the given species name. Two-step lookup: first calls flora_search_sp?scientific_name={name}
        to resolve the plant_id, then calls locs_sp?id={plant_id} for county data.
        Results are the raw passthrough response from the Michigan Flora county API endpoint.
        Cached permanently (no TTL) — Michigan Flora data does not change;
        use ?refresh=true to force a re-fetch from upstream.
      parameters:
        - name: name
          in: query
          required: true
          description: Scientific name to look up county records for (e.g. Asclepias tuberosa)
          schema:
            type: string
            minLength: 1
        - name: refresh
          in: query
          required: false
          description: If true, bypasses cache and fetches fresh from Michigan Flora API
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: County occurrence records (check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MifloraCountiesResponse"
        "400":
          description: Missing or empty name parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Michigan Flora API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /miflora/images:
    get:
      operationId: getMifloraImages
      tags: [miflora]
      summary: Get the full photo gallery for a Michigan Flora species
      description: >
        Returns all images from the Michigan Flora allimage_info endpoint for a given species,
        each enriched with constructed absolute image_url and thumbnail_url fields.
        Two-step lookup: flora_search_sp resolves the plant_id, then allimage_info fetches
        all available photos. Results are cached permanently (no TTL) — Michigan Flora image
        data does not change. Use ?refresh=true to force a re-fetch.
        Image URL formula (reverse-engineered from Michigan Flora frontend):
        Full: https://michiganflora.net/static/species_images/_pid_{plant_id}/{image_id}.jpg
        Thumbnail: https://michiganflora.net/static/species_images/_pid_{plant_id}/thumb_{image_id}.jpg
      parameters:
        - name: name
          in: query
          required: true
          description: Scientific name to look up (e.g. Asclepias tuberosa)
          schema:
            type: string
            minLength: 1
        - name: refresh
          in: query
          required: false
          description: If true, bypasses cache and fetches fresh from Michigan Flora API
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Image gallery (check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MifloraImagesResponse"
        "400":
          description: Missing or empty name parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Michigan Flora API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
    FernsProvenance:
      type: object
      description: >
        Provenance block present on every FERNS API response.
        Identity fields (source_id, fetched_at, method, upstream_url) are always present.
        Text fields (general_summary, technical_details) are conditionally present
        based on the provenance_verbosity query parameter (full|summary|none).
      required:
        - source_id
        - fetched_at
        - method
        - upstream_url
      properties:
        source_id:
          type: string
          description: Stable identifier for this data source (e.g. bonap-napa)
        fetched_at:
          type: string
          format: date-time
          description: When this record was obtained from the source
        method:
          type: string
          description: "How the data was obtained: api_fetch | blob_import | llm_synthesis"
        upstream_url:
          type: string
          description: Where this data came from (API endpoint, file path, or registry entry)
        general_summary:
          type: string
          description: Plain language description readable by a homeowner or community member
        technical_details:
          type: string
          description: >
            Research-grade description: methods, measurement protocols, algorithms,
            citations, and transformations — sufficient for a scientist to evaluate and reproduce
        matched_input:
          type: string
          description: >
            The normalized input that was actually used for this lookup (e.g., the name
            as queried). Present on endpoints that accept a name parameter.
...
    MifloraSpeciesResponse:
      type: object
      description: >
        FERNS envelope for Michigan Flora species lookup.
        data contains the passthrough response from Michigan Flora with one additive
        enrichment: pimage_info is augmented with image_url and thumbnail_url fields
        constructed from the plant_id and image_id using the Michigan Flora static
        asset URL formula. All other source fields are returned unchanged.
      required:
        - source_url
        - found
        - cache_status
        - queried_at
        - provenance
      properties:
        source_url:
          type: string
          nullable: true
          description: >
            Michigan Flora species page URL (https://michiganflora.net/species/{plant_id}).
            Null when species not found.
        found:
          type: boolean
          description: Whether the species was found in Michigan Flora
        cache_status:
          type: string
          enum: [hit, miss, error]
        queried_at:
          type: string
          format: date-time
        data:
          nullable: true
          description: >
            Passthrough response from Michigan Flora. Contains search_records (array from
            flora_search_sp — may include subspecies and varieties), spec_text (taxonomic
            details and description), synonyms (raw response — either {synonyms:[...]} or
            {message:'No synonyms found'}), and pimage_info (primary image metadata).
            pimage_info is additively enriched by FERNS with image_url and thumbnail_url
            (constructed absolute URLs). All other source fields are returned unchanged.
            The st field uses the literal string 'NULL' for unknown/absent status.
            The c field is always a string; '*' means adventive (non-native).
            Null when found is false.
          type: object
          additionalProperties: true
        provenance:
          $ref: "#/components/schemas/FernsProvenance"
    MifloraCountiesResponse:
      type: object
      description: >
        FERNS envelope for Michigan Flora county occurrence data.
        data is the raw passthrough response from the Michigan Flora county endpoint.
      required:
        - source_url
        - found
        - cache_status
        - queried_at
        - provenance
      properties:
        source_url:
          type: string
          nullable: true
          description: Michigan Flora species page URL for this species.
        found:
          type: boolean
          description: Whether county records were found for this species
        cache_status:
          type: string
          enum: [hit, miss, error]
        queried_at:
          type: string
          format: date-time
        data:
          nullable: true
          description: >
            Raw passthrough response from the Michigan Flora county API endpoint.
            Array of county records for all 83 Michigan counties. Each record includes
            county name, FIPS code, and occurrence status. All source fields unchanged.
        provenance:
          $ref: "#/components/schemas/FernsProvenance"
    MifloraImageRecord:
      type: object
      description: >
        A single image record from the Michigan Flora allimage_info endpoint,
        enriched with constructed absolute image_url and thumbnail_url fields.
      required:
        - image_id
        - image_url
        - thumbnail_url
      properties:
        image_id:
          description: Michigan Flora internal image ID
          oneOf:
            - type: integer
            - type: string
        image_name:
          type: string
          nullable: true
          description: Short descriptive name for the image (e.g. "flowers", "fruit")
        caption:
          type: string
          nullable: true
          description: Caption text from Michigan Flora for this image
        photographer:
          type: string
          nullable: true
          description: Photographer credit for this image
        image_url:
          type: string
          description: >
            Absolute URL to the full-size image on Michigan Flora's server.
            Constructed as: https://michiganflora.net/static/species_images/_pid_{plant_id}/{image_id}.jpg
        thumbnail_url:
          type: string
          description: >
            Absolute URL to the thumbnail image on Michigan Flora's server.
            Constructed as: https://michiganflora.net/static/species_images/_pid_{plant_id}/thumb_{image_id}.jpg
    MifloraImagesResponse:
      type: object
      description: >
        FERNS envelope for Michigan Flora image gallery.
        data is the array of image records from allimage_info, each enriched with
        constructed absolute image_url and thumbnail_url. Cached permanently (no TTL).
      required:
        - source_url
        - found
        - cache_status
        - queried_at
        - provenance
      properties:
        source_url:
          type: string
          nullable: true
          description: Michigan Flora species page URL. Null when species not found.
        found:
          type: boolean
          description: Whether images were found for this species in Michigan Flora
        cache_status:
          type: string
          enum: [hit, miss, error]
        queried_at:
          type: string
          format: date-time
        data:
          nullable: true
          description: >
            Array of image records from Michigan Flora allimage_info, each enriched with
            image_url and thumbnail_url. Null when found is false.
          type: array
          items:
            $ref: "#/components/schemas/MifloraImageRecord"
        provenance:
          $ref: "#/components/schemas/FernsProvenance"
...